+++ app/js/controllers/user_controller.js
@@ -1,10 +1,11 @@
 module.exports = function(app) {
   app.controller('UserController', ['$scope', '$http', 'Resource', '$stateParams',
     function($scope, $http, Resource, $stateParams) {
-      $scope.pending = [];
-      $scope.approved = [];
+      $scope.myChallenges = [];
+      $scope.favorites = [];
       $scope.newPost = {}
       $scope.challengeService = new Resource('/challenges');
+      $scope.favoriteService = new Resource('/favorites');
 
       $scope.submitChallenge = function() {
         // Set the current date and time
@@ -20,13 +21,20 @@ module.exports = function(app) {
           if (err) return console.log(err);
         });
       }
-
-
-
-
-
-
-
-
+      $scope.getAllFavorites = function() {
+        $scope.favoriteService.getAll((err, res) => {
+          if (err) return console.log(err);
+          $scope.favorites = res;
+        });
+      };
+      $scope.getUserChallenges = function() {
+        $http.get('http://localhost:3000/api/mychallenges')
+          .then((res) => {
+            console.log(res);
+            $scope.myChallenges = res.data;
+          }, (err) => {
+            console.log(err);
+          });
+      };
   }]);
 };
