+++ routes/user_routes.js
@@ -14,80 +14,10 @@ userRouter.get('/currentuser', jwtAuth, jsonParser, (req, res) => {
   });
 });
 
-// Add new favorite if ids are valid and the user hasn't already favorited
-// this challenge
-userRouter.post('/favorites', jwtAuth, jsonParser, (req, res) => {
-  User.findOne({ _id: req.body.userId }).exec()
-    .then((data) => {
-      if (!data) throw new Error('invalid user');
-      return Challenge.findOne({ _id: req.body.challengeId }).exec();
-    })
-    .then((data) => {
-      if (!data) throw new Error('invalid challenge');
-      return Favorite.findOne(req.body).exec();
-    })
-    .then((data) => {
-      if (data) throw new Error('already favorited');
-      return Favorite.create(req.body);
-    })
-    .then((data) => {
-      res.status(200).json(data);
-    })
-    .catch((err) => {
-      // Check if this was an error we threw instead of a
-      // mongoose promise rejection
-      if (err instanceof Error) {
-        return res.status(400).json({ msg: err.message });
-      }
-      handleDBError(err, res);
-    });
-});
-
-// Retrieves all challenges favorited by the user
-userRouter.get('/favorites', jwtAuth, jsonParser, (req, res) => {
-  Favorite.find({ userId: req.user._id }).exec()
-    .then((favorites) => {
-      const favIds = favorites.map((fav) => fav.challengeId);
-      return Challenge.find({ _id: { $in: favIds } }).exec();
-    })
-    .then((favChallenges) => res.status(200).json(favChallenges))
-    .catch((err) => handleDBError(err, res));
-});
-
-userRouter.post('/favorites', jwtAuth, jsonParser, (req, res) => {
-  var newFavorite = new Favorite(req.body);
-  newFavorite.save((err, data) => {
-    if (err) return handleDBError(err, res);
-    res.status(200).json(data);
-  });
-});
-
-userRouter.delete('/favorites/:id', jwtAuth, jsonParser, (req, res) => {
-  Favorite.remove({ challengeId: req.params.id }, (err) => {
-    if (err) return handleDBError(err, res);
-    res.status(200).json({ msg: 'Successfully Deleted Challenge' });
-  });
-});
-
-
 userRouter.get('/mychallenges', jwtAuth, jsonParser, (req, res) => {
-  Challenge.find({ _id: req.user._id }, (err, data) => {
+  Challenge.find({ userId: req.user._id }, (err, data) => {
     if (err) return handleDBError(err, res);
     res.status(200).json(data);
   });
 });
 
-
-userRouter.get('/pending', jwtAuth, jsonParser, (req, res) => {
-  Challenge.find({ _id: req.user._id, isPublished: false }, (err, data) => {
-    if (err) return handleDBError(err, res);
-    res.status(200).json(data);
-  });
-});
-
-userRouter.get('/approved', jwtAuth, jsonParser, (req, res) => {
-  Challenge.find({ _id: req.user._id, isPublished: true }, (err, data) => {
-    if (err) return handleDBError(err, res);
-    res.status(200).json(data);
-  });
-});
